requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
pandas
numpy
scikit-learn
matplotlib
flask
python-dotenv
google-generativeai
